+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add option to always prefer links when downloading from pages that match a regex list of URLs (#100)
 * Add notification option for when a link is downloaded instead of the source
+* Add option to close tabs marked for saving (FF, #68)
 
 # 3.3.0